requirements.txt
flask
requests
openai
psycopg2-binary
python-dotenv
ollama
openbb